feat: decouple temp file creation from execution
We need to create temp files not tied to an execution block so we must
split withTempFile() in two parts.
